Struct easy_fs::bitmap::Fat

pub struct Fat<T> {
    pub fat_cache_mgr: Arc<Mutex<T>>,
    start_block_id: usize,
    byts_per_sec: usize,
    tot_ent: usize,
    vacant_clus: Mutex<VecDeque<u32>>,
    hint: Mutex<usize>,
}

In-memory data structure In FAT32, there are 2 FATs by default. We use ONLY the first one.

Fields
fat_cache_mgr: Arc<Mutex<T>>start_block_id: usize

The first block id of FAT. In FAT32, this is equal to bpb.rsvd_sec_cnt

byts_per_sec: usize

size fo sector in bytes copied from BPB

tot_ent: usize

The total number of FAT entries

vacant_clus: Mutex<VecDeque<u32>>

The queue used to store known vacant clusters

hint: Mutex<usize>

The final unused clus id we found

Implementations

impl<T: CacheManager> Fat<T>

fn get_eight_blk(&self, start: u32) -> Vec<usize>

pub fn get_next_clus_num(
&self,
start: u32,
block_device: &Arc<dyn BlockDevice>
) -> u32

Get the next cluster number pointed by current fat entry.

pub fn get_all_clus_num(
&self,
start: u32,
block_device: &Arc<dyn BlockDevice>
) -> Vec<u32>

In theory, there may also be one function that only reads the first parts, or the needed FAT entries of the file.

pub fn new(
rsvd_sec_cnt: usize,
byts_per_sec: usize,
clus: usize,
fat_cache_mgr: Arc<Mutex<T>>
) -> Self

Create a new FAT object in memory.

Argument

rsvd_sec_cnt: size of BPB

byts_per_sec: literal meaning

clus: the total numebr of FAT entries

pub fn this_fat_inner_cache_num(&self, n: u32) -> usize

Given any valid cluster number N, where in the FAT(s) is the entry for that cluster number Return the sector number of the FAT sector that contains the entry for cluster N in the first FAT

pub fn this_fat_inner_sec_num(&self, n: u32) -> usize

Given any valid cluster number N, where in the FAT(s) is the entry for that cluster number Return the sector number of the FAT sector that contains the entry for cluster N in the first FAT

pub fn this_fat_sec_num(&self, n: u32) -> usize

Given any valid cluster number N, where in the FAT(s) is the entry for that cluster number Return the sector number of the FAT sector that contains the entry for cluster N in the first FAT

pub fn this_fat_ent_offset(&self, n: u32) -> usize

Return the offset (measured by bytes) of the entry from the first bit of the sector of the n is the ordinal number of the cluster

fn set_next_clus(
&self,
block_device: &Arc<dyn BlockDevice>,
current: u32,
next: u32
)

Assign the cluster entry to current to next

pub fn cnt_all_fat(&self, block_device: &Arc<dyn BlockDevice>) -> usize

pub fn alloc_mult(
&self,
block_device: &Arc<dyn BlockDevice>,
alloc_num: usize,
attach: Option<u32>
) -> Vec<u32>

Allocate as many clusters (but not greater than alloc_num) as possible.

pub fn alloc_one(
&self,
block_device: &Arc<dyn BlockDevice>,
attach: Option<u32>
) -> Option<u32>

Find and allocate an cluster from data area. block_device: The target block_device cache_mgr: The cache manager attach: The preceding cluster of the one to be allocated

fn alloc_one_no_attach_locked(
&self,
block_device: &Arc<dyn BlockDevice>
) -> Option<u32>

fn get_next_free_clus(
&self,
start: u32,
block_device: &Arc<dyn BlockDevice>
) -> Option<u32>

pub fn dealloc(&self, block_device: &Arc<dyn BlockDevice>, bit: u32)

Find and allocate an empty block from data area. This function must be changed into a cluster-based one in the future.

pub fn mult_dealloc(&self, block_device: &Arc<dyn BlockDevice>, v: Vec<u32>)

pub fn maximum(&self) -> usize