Union easy_fs::layout::FATDirEnt

pub union FATDirEnt {
    pub short_entry: FATShortDirEnt,
    pub long_entry: FATLongDirEnt,
}

Fields
short_entry: FATShortDirEntlong_entry: FATLongDirEnt
Implementations

impl FATDirEnt

pub fn gen_short_name_numtail(v: Vec<FATDirEnt>, name_res: &mut [u8; 11])

Test whether self is a short entry and whether the short entry name of self is the same type of prefix.

pub fn gen_short_name_prefix(s: String) -> String

Embedded spaces within a long name are allowed. Leading and trailing spaces in a long name are ignored. Leading and embedded periods are allowed in a name and are stored in the long name. Trailing periods are ignored. No ‘~’ or trailing numbers

pub fn get_ord(&self) -> usize

pub fn empty() -> Self

pub fn unused_not_last_entry() -> Self

pub fn unused_and_last_entry() -> Self

pub fn as_bytes(&self) -> &[u8]

pub fn as_bytes_mut(&mut self) -> &mut [u8]

pub fn is_last_long_dir_ent(&self) -> bool

pub fn ord(&self) -> usize

pub fn set_size(&mut self, size: u32)

pub fn get_fst_clus(&self) -> u32

pub fn set_fst_clus(&mut self, fst_clus: u32)

pub fn is_long(&self) -> bool

pub fn is_short(&self) -> bool

pub fn get_short_ent(&self) -> Option<&FATShortDirEnt>

pub fn get_long_ent(&self) -> Option<&FATLongDirEnt>

pub fn get_name(&self) -> String

pub fn get_short_name_array(&self) -> [u8; 11]

pub fn unused(&self) -> bool

pub fn unused_not_last(&self) -> bool

pub fn last_and_unused(&self) -> bool