Struct easy_fs::EasyFileSystem

pub struct EasyFileSystem<T: CacheManager, F: CacheManager> {
    used_marker: PhantomData<T>,
    pub block_device: Arc<dyn BlockDevice>,
    pub fat: Fat<F>,
    pub data_area_start_block: u32,
    pub root_clus: u32,
    pub sec_per_clus: u8,
    pub byts_per_sec: u16,
    ino_cnt: Mutex<u64>,
}

Fields
used_marker: PhantomData<T>block_device: Arc<dyn BlockDevice>

Partition/Device the FAT32 is hosted on.

fat: Fat<F>

FAT information

data_area_start_block: u32

The first data sector beyond the root directory

root_clus: u32

This is set to the cluster number of the first cluster of the root directory, usually 2 but not required to be 2.

sec_per_clus: u8

sector per cluster, usually 8 for SD card

byts_per_sec: u16

Bytes per sector, 512 for SD card

ino_cnt: Mutex<u64>

“New” inode number count

Implementations

impl<T: CacheManager, F: CacheManager> EasyFileSystem<T, F>

pub fn this_fat_ent_offset(&self, n: u32) -> u32

pub fn this_fat_sec_num(&self, n: u32) -> u32

pub fn get_next_clus_num(&self, result: u32) -> u32

impl<T: CacheManager, F: CacheManager> EasyFileSystem<T, F>

pub fn first_data_sector(&self) -> u32

pub fn clus_size(&self) -> u32

impl<T: CacheManager, F: CacheManager> EasyFileSystem<T, F>

pub fn alloc_new_inode(&self) -> u64

pub fn first_sector_of_cluster(&self, n: u32) -> u32

n is the ordinal number of the cluster.

pub fn in_cluster(&self, block_id: u32) -> u32

pub fn open(
block_device: Arc<dyn BlockDevice>,
bpb_fat_cache_mgr: Arc<Mutex<F>>
) -> Arc<Self>

Open the filesystem object.

pub fn root_inode(efs: &Arc<Self>) -> Arc<Inode<T, F>>

Open the root directory

pub fn get_disk_fat_pos(&self, n: u32) -> (u32, usize)

Look up the first sector denoted by inode_id Inode is not natively supported in FAT32. However, fst_clus may be used as the inode_id Only path is an UNIQUE id to a file in FAT32.