Struct easy_fs::Inode

pub struct Inode<T: CacheManager, F: CacheManager> {
    pub file_content: Mutex<FileContent<T>>,
    pub file_type: ReadOnly<DiskInodeType>,
    pub parent_dir: Option<(Arc<Self>, u32)>,
    pub fs: Arc<EasyFileSystem<T, F>>,
    pub time: Mutex<InodeTime>,
}

The functionality of ClusLi & Inode can be merged. The struct for file information

Fields
file_content: Mutex<FileContent<T>>

File Content

file_type: ReadOnly<DiskInodeType>

File type

parent_dir: Option<(Arc<Self>, u32)>

The parent directory of this inode

fs: Arc<EasyFileSystem<T, F>>

file system

time: Mutex<InodeTime>

Struct to hold time related information

Implementations

impl<T: CacheManager, F: CacheManager> Inode<T, F>

Constructor

pub fn new(
fst_clus: u32,
file_type: DiskInodeType,
size: Option<u32>,
parent_dir: Option<(Arc<Self>, u32)>,
fs: Arc<EasyFileSystem<T, F>>
) -> Arc<Self>

Constructor for Inodes

Arguments

fst_clus: The first cluster of the file type_: The type of the inode determined by the file size: NOTE: the size field should be set to None for a directory parent_dir: parent directory fs: The pointer to the file system

impl<T: CacheManager, F: CacheManager> Inode<T, F>

Basic Funtions

pub fn get_first_clus(&self) -> Option<u32>

Get first cluster of inode. If cluster list is empty, it will return None.

Warning

This function is an external interface. May cause DEADLOCK if called on internal interface

pub fn get_inode_num(&self) -> Option<u32>

Get inode number of inode. See first sector number as inode number.

pub fn get_neighboring_sec(&self, inner_cache_id: usize) -> Vec<usize>

!!! This function have many bugs Get the neighboring 8 or fewer(the trailing mod-of-eight blocks of the file) blocks of inner_block_id,

pub fn is_dir(&self) -> bool

Check if file type is directory

pub fn is_file(&self) -> bool

Check if file type is file

pub fn get_file_size(&self) -> u32

Get file size

Warning

This function is an external interface. May cause DEADLOCK if called on internal interface

pub fn get_file_clus(&self) -> u32

Get the number of clusters corresponding to the size.

fn total_clus(&self, size: u32) -> u32

Get the number of clusters needed after rounding up according to size.

fn get_block_id(
&self,
lock: &MutexGuard<'_, FileContent<T>>,
blk: u32
) -> Option<u32>

Get block id corresponding to the blk

impl<T: CacheManager, F: CacheManager> Inode<T, F>

File Content Operation

fn alloc_clus(
&self,
lock: &mut MutexGuard<'_, FileContent<T>>,
alloc_num: usize
)

Allocate required clusters

fn dealloc_clus(
&self,
lock: &mut MutexGuard<'_, FileContent<T>>,
dealloc_num: usize
)

Deallocate required cluster If the required number is greater than the number of clusters in the file, all clusters will be deallocated

pub fn modify_size(
&self,
lock: &mut MutexGuard<'_, FileContent<T>>,
diff: isize
)

Change the size of current file. This operation is ignored if the result size is negative

Warning

This function will lock parent’s file content. May cause DEADLOCK

impl<T: CacheManager, F: CacheManager> Inode<T, F>

Open File Table

pub fn open_tab(file: OpenTabCmd<T, F>) -> Option<Arc<Self>>

Open file table static operation function.

impl<T: CacheManager, F: CacheManager> Inode<T, F>

IO

pub fn read_at_block_cache(
&self,
lock: &mut MutexGuard<'_, FileContent<T>>,
offset: usize,
buf: &mut [u8]
) -> usize

The get_block_cache version of read_at Read the inode(file) denoted by self, starting from offset. read till the minor of buf.len() and self.size

Arguments

buf: The destination buffer of the read data

offset: The offset

block_device: the block_dev

pub fn write_at_block_cache(
&self,
lock: &mut MutexGuard<'_, FileContent<T>>,
offset: usize,
buf: &[u8]
) -> usize

impl<T: CacheManager, F: CacheManager> Inode<T, F>

Directory Operation

fn dir_iter<'a>(
&'a self,
lock: MutexGuard<'a, FileContent<T>>,
offset: Option<u32>,
mode: DirIterMode,
forward: bool
) -> DirIter<'a, T, F>

Iter Construct

fn expand_dir_size(&self, lock: &mut MutexGuard<'_, FileContent<T>>) -> Result

fn alloc_dir_ent<'a>(
parent_dir: &'a Arc<Self>,
lock: MutexGuard<'a, FileContent<T>>,
alloc_num: usize
) -> Result<(u32, MutexGuard<'a, FileContent<T>>), ()>

return the offset of last free entry

impl<T: CacheManager, F: CacheManager> Inode<T, F>

Delete Little thinking: The real delete operation is done by unlink syscall(this maybe a big problem) So we don’t care about atomic deletes in the filesystem We can recycle resources at will, and don’t care about the resource competition of this inode

fn delete_self_dir_ent(&self)

Delete the short and the long entry of self from parent_dir

pub fn delete_from_disk(trash: Arc<Self>) -> Result<(), ()>

Delete the file from the disk, deallocating both the directory entries (whether long or short), and the occupied clusters.

impl<T: CacheManager, F: CacheManager> Inode<T, F>

Create

pub fn create(
parent_dir: &Arc<Self>,
name: String,
file_type: DiskInodeType
) -> Result<Arc<Inode<T, F>>, ()>

Create a file or a directory from the parent.

fn get_long_name_slice(name: &String, long_ent_num: usize) -> [u16; 13]

Construct a [u16,13] corresponding to the long_ent_num’th 13-u16 or shorter name slice NOTE: the first entry is of number 1 for long_ent_num

fn gen_short_name_slice(
parent_dir: &Arc<Self>,
name: &String,
short_name_slice: &mut [u8; 11]
) -> Result<(), ()>

pub fn from_ent(
parent_dir: &Arc<Self>,
ent: &FATShortDirEnt,
offset: u32
) -> Arc<Self>

Create a file from directory entry.

Arguments

parent_dir: the parent directory inode pointer ent: the short entry as the source of information offset: the offset of the short directory entry in the parent_dir

fn write_back_dir_ent(
parent_dir: &Arc<Self>,
short_ent_offset: u32,
lock: MutexGuard<'_, FileContent<T>>,
short_ent: FATShortDirEnt,
long_ents: Vec<FATLongDirEnt>
)

Write back both long and short directories. The short directory is created from the fst_clus and the name.

fn fill_empty_dir(
parent_dir: &Arc<Inode<T, F>>,
current_dir: &Arc<Inode<T, F>>,
current_lock: MutexGuard<'_, FileContent<T>>,
fst_clus: u32
)

Fill out an empty directory with only the ‘.’ & ‘..’ entries.

impl<T: CacheManager, F: CacheManager> Inode<T, F>

pub fn ls(&self) -> Result<Vec<(String, FATShortDirEnt)>, ()>

ls - General Purose file filterer

Argument/Modes

Apart from None mode, all modes will quit IMMEDIATELY returning a vector of one after the first successful match. Modes are conveyed through enum DirFilter. Four modes are provided, namely:

Name(String): Used for exact match of names.

FstClus(u64): First Cluster matching. Note that this shouldn’t be used for zero-sized files for they MAY contain NO by specification. For this mode, it returns an Err(()) for the next of reaching the last item.

DirentBegOffset(u32): Search should begin with the

None: List all files in self.

WARNING

The definition of OFFSET is CHANGED for this item. It should point to the NEXT USED entry whether it as a long entry whenever possible or a short entry if no long ones exist.

Return value

On success, the function returns Ok(_). On failure, multiple chances exist: either the Vec is empty, or the Result is Err(()).

Implementation Information

The iterator stops at the last available item when it reaches the end, returning None from then on, so relying on the offset of the last item to decide whether it has reached an end is not recommended.

pub fn find_local(
&self,
req_name: String
) -> Result<Option<(String, FATShortDirEnt, u32)>, ()>

impl<T: CacheManager, F: CacheManager> Inode<T, F>

pub fn stat(&self) -> (i64, i64, i64, i64, u64)

Return the stat structure to self file.

pub fn dirent_info(
&self,
offset: u32,
length: usize
) -> Result<Vec<(String, usize, u64, FATDiskInodeType)>, ()>

Get a dirent information from the self at offset Return None if self is not a directory.

Argument

offset: u32, the offset within the self directory.