Crate easy_fs

Re-exports

pub use block_cache::Cache;

pub use block_cache::CacheManager;

pub use layout::DataBlock;

pub use layout::DiskInodeType;

Modules

bitmap

block_cache

block_dev

dir_iter

efs

layout

vfs

Structs

EasyFileSystem

Inode

The functionality of ClusLi & Inode can be merged. The struct for file information

Constants

BLOCK_SZ

CACHE_SZ

Traits

BlockDevice

We should regulate the behavior of this trait on FAILURE e.g. What if buf.len()>BLOCK_SZ for read_block? e.g. Does read_block clean the rest part of the block to be zero for buf.len()!=BLOCK_SZ in write_block() & read_block() e.g. What if buf.len()<BLOCK_SZ for write_block?